backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# File Handling
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.1.0

# Geospatial (Optional - can be enabled later)
# geoalchemy2==0.14.2
# shapely==2.0.2
geopy==2.4.0

# 3D Models (Optional)
trimesh==4.0.5
pygltflib==1.16.1

# Cloud Storage (Optional)
boto3==1.34.0

# Background Tasks
redis==5.0.1
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0